[I2C]: Use correct port address types in i2c-velleman.c

+++ drivers/i2c/busses/i2c-velleman.c
@@ -31,7 +31,7 @@ DECL|macro|I2C_CMASK
 mdefine_line|#define I2C_CMASK&t;0xf7
 multiline_comment|/* --- Convenience defines for the parallel port:&t;&t;&t;*/
 DECL|macro|BASE
-mdefine_line|#define BASE&t;(unsigned int)(data)
+mdefine_line|#define BASE&t;(unsigned long)(data)
 DECL|macro|DATA
 mdefine_line|#define DATA&t;BASE&t;&t;&t;/* Centronics data port&t;&t;*/
 DECL|macro|STAT
